requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pdfplumber>=0.10.0
easyocr>=1.7.0
pytesseract>=0.3.10
langdetect>=1.0.9
rapidfuzz>=3.5.0
scikit-image>=0.21.0
diskcache>=5.6.0
pyzbar>=0.1.9
python-magic>=0.4.27
PyYAML>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
